fix errro

+++ experiments/hrnet/utils.py
@@ -98,7 +98,8 @@ def get_pose_point(prediction_logits: torch.Tensor):
 
 @torch.no_grad()
 def compute_pck_metric(prediction, label, tau=0.5):
-    pose_point_pred = get_pose_point(prediction)
+    pose_point_pred = get_pose_point(prediction).float()
+    label = label.float()
     torso_dims = torch.norm(label[:, 1] - label[:, 11], dim=-1)
     distance = torch.norm(pose_point_pred-label, dim=-1)
     return torch.mean((torch.less_equal(distance, torso_dims*tau)) * 1.)
@@ -109,7 +110,8 @@ K_VEC = torch.tensor([.26, .25, .25, .35, .35, .79, .79, .72, .72, .62, .62, 1.0
 
 @torch.no_grad()
 def compute_oks_metric(prediction, label):
-    pose_point_pred = get_pose_point(prediction)
+    pose_point_pred = get_pose_point(prediction).float()
+    label = label.float()
     distance = torch.norm(pose_point_pred - label, dim=-1)
     k_vec = K_VEC.to(prediction.device)
     body_area = torch.prod(label.max(dim=1)[0] - label.min(dim=1)[0], dim=-1).unsqueeze(-1)
